demos/use_cases/orchestrating_agents/pyproject.toml
[tool.poetry]
name = "api-server"
version = "0.1.0"
description = ""
authors = ["Adil Hafeez <info@katanemo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-instrumentation-fastapi = "^0.49b0"
fastapi = "^0.115.4"
pyyaml = "^6.0.2"
uvicorn = "^0.32.0"
opentelemetry-api = "^1.28.0"
opentelemetry-sdk = "^1.28.0"
opentelemetry-exporter-otlp = "^1.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api-server = "api_server.main:app"
